htmldiff2
We waited a while for this but ultimately had to do o friendly fork of
of Armin Ronacher's htmldiff to support another open source project. I
would be happy to deprecate this in favor of the original if it ever
gets updated.

For more context see: https://github.com/mitsuhiko/htmldiff/issues/7